added client side auth util
