+++ client/package.json
@@ -16,6 +16,7 @@
     "@mui/icons-material": "^5.15.11",
     "@mui/material": "^5.15.11",
     "bcrypt": "^5.1.1",
+    "jwt-decode": "^4.0.0",
     "mongoose": "^8.2.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
