+++ client/src/components/Explore/ProductCard.jsx
@@ -1,18 +1,22 @@
-import * as React from "react";
 import { Link } from "react-router-dom";
 import Card from "@mui/material/Card";
 import CardContent from "@mui/material/CardContent";
 import CardMedia from "@mui/material/CardMedia";
 import Typography from "@mui/material/Typography";
 import { Button, CardActionArea, CardActions, Box, Grid } from "@mui/material";
+import Checkbox from '@mui/material/Checkbox';
+import FavoriteBorder from '@mui/icons-material/FavoriteBorder';
+import Favorite from '@mui/icons-material/Favorite';
 import Data from '../../productData.json'; // Sample product data
+import {Tooltip} from "@mui/material";
+
 
 export default function ProductCard() {
   return (
-    <Grid container spacing={2} marginTop={4}>
+    <Grid container spacing={5} marginTop={4}>
       {Data.map((result, index) => (
         <Grid item xs={12} sm={6} md={4} key={index} align="center">
-          <Card sx={{ maxWidth: 275, padding: "10px", marginBottom: '10px' }}>
+          <Card sx={{ maxWidth: 275, padding: "10px" }}>
             <CardActionArea component={Link} to={`/product/${result.id}`}>
               <Box sx={{ height: "100px", width: "100px" }}>
                 <CardMedia
@@ -27,11 +31,13 @@ export default function ProductCard() {
                   }}
                 />
               </Box>
+
               <CardContent align='left'>
                 <Typography gutterBottom variant="h5" component="div">
                   {/* NAME data */}
                   {result.name}
                 </Typography>
+                
                 <Typography variant="body2" color="text.secondary"
                   sx={{
                     // Apply ellipsis and hidden overflow when description content exceeds two lines length
@@ -48,13 +54,24 @@ export default function ProductCard() {
               </CardContent>
             </CardActionArea>
             <CardActions>
-              <Button
-                size="small"
-                variant="contained"
-                sx={{ bgcolor: "secondary.main", color: "primary.main", textTransform: 'none' }}
-              >
-                Add to cart
-              </Button>
+              <Grid container justifyContent="space-between">
+                <Grid item>
+                  <Button
+                    size="small"
+                    variant="contained"
+                    sx={{ bgcolor: "secondary.main", color: "primary.main", textTransform: 'none' }}
+                  >
+                    Add to cart
+                  </Button>
+                </Grid>
+                <Grid item>
+                  <Box>
+                  <Tooltip title="Add to wishlist" placement="right">
+          <Checkbox icon={<FavoriteBorder />} checkedIcon={<Favorite />} />
+          </Tooltip>
+                  </Box>
+                </Grid>
+              </Grid>
             </CardActions>
           </Card>
         </Grid>
