+++ client/src/components/Signup/BuyerSignup.jsx
@@ -7,6 +7,7 @@ import Box from '@mui/material/Box';
 import Typography from '@mui/material/Typography';
 import Container from '@mui/material/Container';
 import ShoppingBasketIcon from '@mui/icons-material/ShoppingBasket';
+import Auth from '../../utils/auth';
 
 
 function Copyright(props) {
