+++ client/src/pages/SingleProductView.jsx
@@ -1,0 +1,9 @@
+import SingleProduct from "../components/Explore/SingleProduct";
+
+// Page that renders when clicking on a single product
+export default function SingleProductView() {
+  
+  return (
+    <SingleProduct/>
+  );
+}
